Trying prevent populateContainers method.
